Move performance warning to vue
