+++ site/app/libraries/Output.php
@@ -202,6 +202,11 @@ HTML;
         $this->addInternalJs('menu.js');
         $this->addInternalJs('testcase-output.js');
         $this->addInternalJs('markdown.js');
+
+        // Add vue support
+        $this->addVendorJs(FileUtils::joinPaths('vue', 'vue.runtime.global.prod.js'));
+        $this->css->add($this->timestampResource('submitty-vue.css', 'mjs/vue'));
+        $this->js->add($this->timestampResource('submitty-vue.umd.js', 'mjs/vue'));
     }
 
     /**
@@ -559,17 +564,6 @@ HTML;
 
     public function addInternalJs($file, $folder = 'js') {
         $this->addJs($this->timestampResource($file, $folder));
-    }
-
-    /**
-     * @param array<mixed>$args
-     */
-    public function renderVue(string $page, array $args = []): string {
-        $this->addVendorJs(FileUtils::joinPaths('vue', 'vue.runtime.global.prod.js'));
-        $this->css->add($this->timestampResource('style.css', 'mjs/vue'));
-        $this->js->add($this->timestampResource('submitty-vue.umd.js', 'mjs/vue'));
-
-        return $this->renderTwigTemplate('Vue.twig', ["page" => $page, "args" => $args]);
     }
 
     public function addVendorJs($file) {
