+++ site/vue/src/components/performanceWarning.vue
@@ -1,0 +1,14 @@
+<template>
+  <div
+    id="performance-warning-system-message"
+    class="system-message warning"
+    style="display: none;"
+  >
+    <span id="performance-warning-system-message-text" />
+    <i
+      id="close-performance-warning-system-message"
+      class="fa fa-times"
+      style="padding: 4px; cursor: pointer;"
+    />
+  </div>
+</template>
