+++ site/vue/src/main.ts
@@ -1,23 +1,19 @@
 import { createApp } from 'vue';
-import Unknown from './pages/Unknown.vue';
+import Unknown from './components/Unknown.vue';
 
 const exports = {
-    async render(target: string, page: string, args: Record<string, unknown> = {}) {
+    async render(target: string | Element, component: string, args: Record<string, unknown> = {}) {
         const root_component = await (async () => {
             try {
-                return (await import(`./pages/${page}.vue`) as { default: Parameters<typeof createApp>[0] }).default;
+                return (await import(`./components/${component}.vue`) as { default: Parameters<typeof createApp>[0] }).default;
             }
             catch (e) {
-                console.error(`Could not render vue page ${page}:`, e);
+                console.error(`Could not find vue component ${component}:`, e);
                 return Unknown;
             }
         })();
 
-        const app = createApp(root_component, { page });
-
-        for (const [key, value] of Object.entries(args)) {
-            app.provide(key, value);
-        }
+        const app = createApp(root_component, { component, ...args });
 
         app.mount(target);
     },
